+++ app/features/showcases-qr-generator/types/type.ts
@@ -1,0 +1,13 @@
+export type PageInformation = {
+  title: string;
+  description: string;
+  language: string;
+};
+
+export type QRCodeFormat = "png" | "jpg";
+
+export type QRCodeOptions = {
+  size: number;
+  level: "L" | "M" | "Q" | "H";
+  marginSize: number;
+};