+++ react-router.config.ts
@@ -6,9 +6,9 @@ export default {
   ssr: true,
   future: {
     v8_middleware: true,
+    v8_splitRouteModules: true,
+    v8_viteEnvironmentApi: true,
     unstable_optimizeDeps: true,
-    unstable_splitRouteModules: true,
     unstable_subResourceIntegrity: true,
-    unstable_viteEnvironmentApi: true,
   },
 } satisfies Config;